requirements.txt
# === Core Deep Learning ===
torch>=2.0.0
torchvision>=0.15.0

# === Manipolazione Dati e Immagini ===
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0

# === Visualizzazione ed EDA ===
matplotlib>=3.5.0
seaborn>=0.11.0

# === Metriche e Valutazione ===
scikit-learn>=1.0.0

# === Web Demo ===
streamlit>=1.20.0

# === Supporto Notebook ===
jupyterlab>=3.0.0
ipykernel>=6.0.0
